+++ src/library/DIPS.Mobile.UI/Components/Pickers/ItemPicker/ItemPicker.cs
@@ -11,7 +11,7 @@ public partial class ItemPicker : ContentView
 {
     private readonly ContextMenu m_contextMenu = new();
     private readonly Chip m_chip = new(){AutomationId = "ItemPickerChip".ToDUIAutomationId<ItemPicker>()};
-
+    
     public ItemPicker()
     {
         m_chip.SetBinding(IsEnabledProperty, static (ItemPicker itemPicker) => itemPicker.IsEnabled, source: this);